Wrote the Windows version of GitBox with support for listing repositories inside GitBox, creating empty repositories, importing existing ones and cloning from GitBox.
Wrote a small setup script for Windows which adds the gitbox batch file to the
user's own PATH environmental variable.

Added some assembly instructions on how to create a distributable bundle for
the program.

Added a license file with the MIT license in it.

Wrote up a very basic (but hopefully complete) user manual.